fixed ArrayListSerializer to add commas between list values
