+++ kafka/cassandraConsumer/src/main/java/CommonConsumer.java
@@ -1,0 +1,38 @@
+import org.apache.kafka.clients.consumer.ConsumerConfig;
+import org.apache.kafka.clients.consumer.KafkaConsumer;
+import org.apache.kafka.common.serialization.StringDeserializer;
+
+import java.util.Arrays;
+import java.util.Properties;
+import java.util.regex.Pattern;
+
+public class CommonConsumer {
+
+    public static Properties setKafkaProperties(String groupId) {
+        String bootstrapServers = "localhost:9092";
+
+        // New consumer configs (Kafka docs)
+        Properties properties = new Properties();
+
+        properties.setProperty(ConsumerConfig.BOOTSTRAP_SERVERS_CONFIG, bootstrapServers);
+        properties.setProperty(ConsumerConfig.KEY_DESERIALIZER_CLASS_CONFIG, StringDeserializer.class.getName());
+        properties.setProperty(ConsumerConfig.VALUE_DESERIALIZER_CLASS_CONFIG, StringDeserializer.class.getName());
+        properties.setProperty(ConsumerConfig.GROUP_ID_CONFIG, groupId);
+        properties.setProperty(ConsumerConfig.AUTO_OFFSET_RESET_CONFIG, "earliest"); // "earliest/latest/none"
+
+        return properties;
+    }
+
+    public KafkaConsumer<String, String> createAndSubscribe(Properties properties, String topic) {
+        // create consumer
+        KafkaConsumer<String, String> consumer =
+                new KafkaConsumer<String, String>(properties);
+
+        Pattern regexP = Pattern.compile("([a-zA-Z0-9]+)(.*)");
+
+        // subscribe consumer to our topic(s)
+        consumer.subscribe(Arrays.asList(topic));
+
+        return consumer;
+    }
+}
