+++ kafka/cassandraConsumer/src/main/resources/anomaly/META-INF/MANIFEST.MF
@@ -1,3 +1,3 @@
 Manifest-Version: 1.0
-Main-Class: AnomalyCassandra
+Main-Class: MovingAvgCassandra
