+++ kafka/streamsKafka/src/main/java/kafkaStreams/ArrayListDeserializer.java
@@ -6,7 +6,6 @@ import java.io.ByteArrayInputStream;
 import java.io.DataInputStream;
 import java.io.IOException;
 import java.util.ArrayList;
-import java.util.Comparator;
 import java.util.Map;
 
 // PriorityQueueDeserializer<T> implements Deserializer<PriorityQueue<T>>
@@ -32,7 +31,7 @@ public class ArrayListDeserializer<T> implements Deserializer<ArrayList<T>> {
             return null;
         }
         final ArrayList<T> arrayList = new ArrayList<>();
-        // i took out comparator from ArrayList<>()... may need to add it elsewhere....
+
         final DataInputStream dataInputStream = new DataInputStream(new ByteArrayInputStream(bytes));
         try {
             final int records = dataInputStream.readInt();
